backend/requirements.txt
Flask==2.2.5
flask-cors==4.0.0
pdfplumber==0.8.0
pandas==2.2.3
mysql-connector-python==8.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
werkzeug==3.0.2
